Moved function tests to test_function.py
